less explicit unwrapping

+++ invesdwin-context-python-parent/invesdwin-context-python-runtime-contract/src/test/java/de/invesdwin/context/python/runtime/contract/InputsAndResultsTestPercent.java
@@ -32,7 +32,7 @@ public class InputsAndResultsTestPercent {
         //putPercentVector
         final Percent[] putPercentVector = new Percent[3];
         for (int i = 0; i < putPercentVector.length; i++) {
-            putPercentVector[i] = new Percent(new Decimal((i + 1) + "." + (i + 1)).doubleValue(), PercentScale.RATE);
+            putPercentVector[i] = new Percent(new Decimal((i + 1) + "." + (i + 1)), PercentScale.RATE);
         }
 
         //putPercentVectorAsList
@@ -43,7 +43,7 @@ public class InputsAndResultsTestPercent {
         for (int row = 0; row < putPercentMatrix.length; row++) {
             final Percent[] vector = new Percent[3];
             for (int col = 0; col < vector.length; col++) {
-                vector[col] = new Percent(new Decimal((row + 1) + "." + (col + 1)).doubleValue(), PercentScale.RATE);
+                vector[col] = new Percent(new Decimal((row + 1) + "." + (col + 1)), PercentScale.RATE);
             }
             putPercentMatrix[row] = vector;
         }
